pyproject.toml
[tool.poetry]
name = "rpg-bot"
version = "0.1.0"
description = "Телеграм бот для RPG игры"
authors = ["Smacktur <smacktur@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.12"
aiogram = "^3.19.0"
sqlalchemy = "^2.0.25"
asyncpg = "^0.29.0"
pydantic = "^2.5.3"
redis = "^5.0.1"
loguru = "^0.7.2"
alembic = "^1.12.1"
celery = "^5.3.6"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.9"
aiosqlite = "^0.19.0"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"